add conditional navigation if logged in and a log out button
