+++ src/App.js
@@ -2,7 +2,8 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import './App.css';
 
-import Navigation from './Components/Navigation';
+import Navigation from './Components/Navigation/Navigation';
+import LoggedInNavigation from "./Components/Navigation/LoggedInNavigation"
 import Home from './Components/Home';
 import Footer from './Components/Footer';
 
@@ -18,30 +19,59 @@ import EditandoTemas from './Components/EditandoTemas';
 import Login from './Components/Login';
 import OldTestament from './Components/EnglishBible/OldTestament';
 import NewTestament from './Components/EnglishBible/NewTestament';
+import LogOut from './Components/LogOut';
 
-function App() {
-  
-  return (
-    <div className="App">
-      <Navigation />
-      <Switch>
-        <Route exact path="/" component={Home} />
-        <Route path="/oldtestament" component={OldTestament} />
-        <Route path="/newtestament" component={NewTestament} />
 
-        <Route path="/antiguotestamento" component={Antiguo} />
-        <Route path="/nuevotestamento" component={Nuevo} />
 
-        <Route path="/temas" component={Temas} />
-        <Route path="/addtemas" component={AddTemas} />
-        <Route path="/borrartemas" component={DeleteTemas} />
-        <Route path="/editartemas" component={EditTemas} />
-        <Route path="/login" component={Login} />
-        <Route path="/editandotemas/:id" component={EditandoTemas} />
-      </Switch>
-      <Footer />
-    </div>
-  );
+export default function App() {
+
+
+  if (!localStorage.getItem('token')) {
+    return (
+      <div className="App">
+        <Navigation />
+        <Switch>
+          <Route exact path="/" component={Home} />
+          <Route path="/oldtestament" component={OldTestament} />
+          <Route path="/newtestament" component={NewTestament} />
+          <Route path="/antiguotestamento" component={Antiguo} />
+          <Route path="/nuevotestamento" component={Nuevo} />
+          <Route path="/temas" component={Temas} />
+          <Route path="/addtemas" component={AddTemas} />
+          <Route path="/borrartemas" component={DeleteTemas} />
+          <Route path="/editartemas" component={EditTemas} />
+          <Route path="/login" component={Login} />
+          <Route path="/editandotemas/:id" component={EditandoTemas} />
+          <Route Path="/logout" component={LogOut} />
+        </Switch>
+        <Footer />
+      </div>
+    );
+  } else if (localStorage.getItem("token")) {
+
+    return (
+      <div className="App">
+        <Navigation />
+        <LoggedInNavigation />
+        <Switch>
+          <Route exact path="/" component={Home} />
+          <Route path="/oldtestament" component={OldTestament} />
+          <Route path="/newtestament" component={NewTestament} />
+          <Route path="/antiguotestamento" component={Antiguo} />
+          <Route path="/nuevotestamento" component={Nuevo} />
+          <Route path="/temas" component={Temas} />
+          <Route path="/addtemas" component={AddTemas} />
+          <Route path="/borrartemas" component={DeleteTemas} />
+          <Route path="/editartemas" component={EditTemas} />
+          <Route path="/login" component={Login} />
+          <Route path="/editandotemas/:id" component={EditandoTemas} />
+          <Route Path="/logout" component={LogOut} />
+        </Switch>
+        <Footer />
+      </div>
+    );
+  }
 }
 
-export default App;
+
+
