+++ src/Components/DeleteTemas.js
@@ -3,7 +3,6 @@ import axios from "axios";
 import "../App.css";
 import * as ReactBootStrap from "react-bootstrap";
 
-import {Login, HeaderButtons} from '../styles2/AddTemasStyles'
 
 import ClipLoader from "react-spinners/ClipLoader";
 import { css } from "@emotion/core";
@@ -52,18 +51,7 @@ function DeleteTemas(props) {
   return (
     <div className="body">
       <div className="container">
-        <HeaderButtons>
-          <Login onClick={() => props.history.push("/addtemas")}>
-            Anadir un tema
-          </Login>
-          <Login onClick={() => props.history.push("/borrartemas")}>
-            Borrar un Tema
-          </Login>
-          <Login onClick={() => props.history.push("/editartemas")}>
-            {" "}
-            Editar un Tema
-          </Login>
-        </HeaderButtons>
+
         <h1 className="display-4 my3">
           <span className="text-dark"> </span> Cuidado cuando borrando Temas
         </h1>
