+++ src/Components/LogOut.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function LogOut() {
+    return (
+        <div>
+            <h1>You have logged out!</h1>
+        </div>
+    )
+}
